MODFISTO-61 budget/ledger unavailable shouldn't include overEncumbered amounts
- moved test for encumbrances to separate class
- update calculations
