+++ src/main/resources/data/budgets/GIFTS-ONE-TIME-FY20.json
@@ -4,7 +4,7 @@
   "budgetStatus": "Active",
   "allowableEncumbrance": 110,
   "allowableExpenditure": 100,
-  "allocated": 28000.00,
+  "allocated": 0.00,
   "fundId": "6506b79b-7702-48b2-9774-a1c538fdd34e",
   "fiscalYearId": "684b5dc5-92f6-4db7-b996-b549d88f5e4e"
 }
