+++ src/test/java/org/folio/rest/impl/TransactionTest.java
@@ -3,40 +3,25 @@ package org.folio.rest.impl;
 import static io.restassured.RestAssured.given;
 import static org.folio.rest.RestVerticle.OKAPI_HEADER_TENANT;
 import static org.folio.rest.impl.StorageTestSuite.storageUrl;
-import static org.folio.rest.impl.TransactionsSummariesTest.ORDER_TRANSACTION_SUMMARIES_ENDPOINT;
 import static org.folio.rest.persist.HelperUtils.getEndpoint;
-import static org.folio.rest.transaction.AllOrNothingHandler.BUDGET_IS_INACTIVE;
-import static org.folio.rest.transaction.AllOrNothingHandler.BUDGET_NOT_FOUND_FOR_TRANSACTION;
-import static org.folio.rest.transaction.AllOrNothingHandler.FUND_CANNOT_BE_PAID;
 import static org.folio.rest.utils.TenantApiTestUtil.deleteTenant;
 import static org.folio.rest.utils.TenantApiTestUtil.prepareTenant;
 import static org.folio.rest.utils.TestEntities.BUDGET;
-import static org.folio.rest.utils.TestEntities.FISCAL_YEAR;
-import static org.folio.rest.utils.TestEntities.FUND;
 import static org.folio.rest.utils.TestEntities.GROUP_FUND_FY;
 import static org.folio.rest.utils.TestEntities.LEDGER;
 import static org.folio.rest.utils.TestEntities.TRANSACTION;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.hasSize;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.math.BigDecimal;
 import java.net.MalformedURLException;
-import java.util.UUID;
 
 import org.apache.commons.lang3.StringUtils;
 import org.folio.rest.jaxrs.model.Budget;
 import org.folio.rest.jaxrs.model.BudgetCollection;
-import org.folio.rest.jaxrs.model.Encumbrance;
-import org.folio.rest.jaxrs.model.Errors;
-import org.folio.rest.jaxrs.model.FiscalYear;
-import org.folio.rest.jaxrs.model.Fund;
 import org.folio.rest.jaxrs.model.Ledger;
 import org.folio.rest.jaxrs.model.LedgerCollection;
 import org.folio.rest.jaxrs.model.LedgerFY;
 import org.folio.rest.jaxrs.model.LedgerFYCollection;
-import org.folio.rest.jaxrs.model.OrderTransactionSummary;
 import org.folio.rest.jaxrs.model.Transaction;
 import org.folio.rest.jaxrs.resource.FinanceStorageLedgerFiscalYears;
 import org.junit.jupiter.api.AfterEach;
@@ -56,9 +41,7 @@ class TransactionTest extends TestBase {
   private static final String LEDGER_QUERY = "?query=fund.id==%s";
   private static final String LEDGER_FY_QUERY = "?query=ledgerId==%s AND fiscalYearId==%s";
   public static final String ALLOCATION_SAMPLE = "data/transactions/allocations/allocation_AFRICAHIST-FY20_ANZHIST-FY20.json";
-  public static final String ENCUMBRANCE_SAMPLE = "data/transactions/encumbrances/encumbrance_AFRICAHIST_306857_1.json";
-  public static final String RESTRICTED_ENCUMBRANCE_SAMPLE = "data/transactions/encumbrances/encumbrance_UNIV-SUBN_312325_1.json";
-  public static final String BUDGET_SAMPLE = "data/budgets/AFRICAHIST-FY20.json";
+
   static String BUDGETS_QUERY = BUDGET.getEndpoint() + FY_FUND_QUERY;
   private static String LEDGERS_QUERY = LEDGER.getEndpoint() + LEDGER_QUERY;
   public static String LEDGER_FYS_ENDPOINT = getEndpoint(FinanceStorageLedgerFiscalYears.class) + LEDGER_FY_QUERY;
@@ -347,522 +330,6 @@ class TransactionTest extends TestBase {
 
   }
 
-
-  @Test
-  void testCreateEncumbranceAllOrNothingIdempotent() throws MalformedURLException {
-
-    String orderId = UUID.randomUUID().toString();
-    createOrderSummary(orderId, 2);
-    JsonObject jsonTx = new JsonObject(getFile(ENCUMBRANCE_SAMPLE));
-    jsonTx.remove("id");
-
-    Transaction encumbrance1 = jsonTx.mapTo(Transaction.class);
-    encumbrance1.getEncumbrance().setSourcePurchaseOrderId(orderId);
-
-    Transaction encumbrance2 = jsonTx.mapTo(Transaction.class);
-    encumbrance2.getEncumbrance().setSourcePurchaseOrderId(orderId);
-    encumbrance2.getEncumbrance().setSourcePoLineId(UUID.randomUUID().toString());
-    String fY = encumbrance1.getFiscalYearId();
-    String fromFundId = encumbrance1.getFromFundId();
-
-
-    // prepare budget/ledger queries
-    String fromBudgetEndpointWithQueryParams = String.format(BUDGETS_QUERY, fY, fromFundId);
-    String fromLedgerEndpointWithQueryParams = String.format(LEDGERS_QUERY, fromFundId);
-
-
-    Budget fromBudgetBefore = getBudgetAndValidate(fromBudgetEndpointWithQueryParams);
-    Ledger fromLedger = getLedgerAndValidate(fromLedgerEndpointWithQueryParams);
-
-    String fromLedgerFYEndpointWithQueryParams = String.format(LEDGER_FYS_ENDPOINT, fromLedger.getId(), fY);
-    LedgerFY fromLedgerFYBefore = getLedgerFYAndValidate(fromLedgerFYEndpointWithQueryParams);
-
-    // create 1st Encumbrance, expected number is 2
-    String encumbrance1Id = postData(TRANSACTION_ENDPOINT, JsonObject.mapFrom(encumbrance1).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201)
-      .extract()
-      .as(Transaction.class).getId();
-
-    // encumbrance do not appear in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), encumbrance1Id, TRANSACTION_TENANT_HEADER).then().statusCode(404);
-
-    // create 2nd Encumbrance
-    String encumbrance2Id = postData(TRANSACTION_ENDPOINT, JsonObject.mapFrom(encumbrance2).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201)
-      .extract()
-      .as(Transaction.class).getId();
-
-    // 2 encumbrances appear in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), encumbrance1Id, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-    getDataById(TRANSACTION.getEndpointWithId(), encumbrance2Id, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-
-    Budget fromBudgetAfter = getBudgetAndValidate(fromBudgetEndpointWithQueryParams);
-    LedgerFY fromLedgerFYAfter = getLedgerFYAndValidate(fromLedgerFYEndpointWithQueryParams);
-
-    // check source budget and ledger totals
-    final double amount = sumValues(encumbrance1.getAmount(), encumbrance2.getAmount());
-    double expectedBudgetsAvailable;
-    double expectedBudgetsUnavailable;
-    double expectedBudgetsEncumbered;
-
-    double expectedLedgersAvailable;
-    double expectedLedgersUnavailable;
-
-
-    expectedBudgetsEncumbered = sumValues(fromBudgetBefore.getEncumbered(), amount);
-    expectedBudgetsAvailable = subtractValues(fromBudgetBefore.getAvailable(), amount);
-    expectedBudgetsUnavailable = sumValues(fromBudgetBefore.getUnavailable(), amount);
-
-    expectedLedgersAvailable = subtractValues(fromLedgerFYBefore.getAvailable(), amount);
-    expectedLedgersUnavailable = sumValues(fromLedgerFYBefore.getUnavailable(), amount);
-
-    assertEquals(expectedBudgetsEncumbered, fromBudgetAfter.getEncumbered());
-    assertEquals(expectedBudgetsAvailable , fromBudgetAfter.getAvailable());
-    assertEquals(expectedBudgetsUnavailable, fromBudgetAfter.getUnavailable());
-
-    assertEquals(expectedLedgersAvailable, fromLedgerFYAfter.getAvailable());
-    assertEquals(expectedLedgersUnavailable , fromLedgerFYAfter.getUnavailable());
-
-
-    //create same encumbrances again
-    postData(TRANSACTION_ENDPOINT, JsonObject.mapFrom(encumbrance1).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201)
-      .extract()
-      .as(Transaction.class);
-
-    postData(TRANSACTION_ENDPOINT, JsonObject.mapFrom(encumbrance2).encodePrettily(), TRANSACTION_TENANT_HEADER)
-      .then()
-      .statusCode(201);
-
-    // check source budget and ledger totals not changed
-    assertEquals(expectedBudgetsEncumbered, fromBudgetAfter.getEncumbered());
-    assertEquals(expectedBudgetsAvailable , fromBudgetAfter.getAvailable());
-    assertEquals(expectedBudgetsUnavailable, fromBudgetAfter.getUnavailable());
-
-    assertEquals(expectedLedgersAvailable, fromLedgerFYAfter.getAvailable());
-    assertEquals(expectedLedgersUnavailable , fromLedgerFYAfter.getUnavailable());
-
-  }
-
-  @Test
-  void testCreateEncumbranceWithNotEnoughBudgetMoney() throws MalformedURLException {
-
-    verifyCollectionQuantity(TRANSACTION_ENDPOINT, TRANSACTION.getInitialQuantity(), TRANSACTION_TENANT_HEADER);
-    String orderId = UUID.randomUUID().toString();
-    createOrderSummary(orderId, 1);
-    JsonObject jsonTx = new JsonObject(getFile(RESTRICTED_ENCUMBRANCE_SAMPLE));
-    jsonTx.remove("id");
-
-    Transaction encumbrance = jsonTx.mapTo(Transaction.class);
-    encumbrance.setAmount(1000000d);
-    encumbrance.getEncumbrance().setSourcePurchaseOrderId(orderId);
-
-    // create Encumbrance
-    postData(TRANSACTION_ENDPOINT, JsonObject.mapFrom(encumbrance).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-      .statusCode(400)
-      .body(containsString(FUND_CANNOT_BE_PAID));
-
-  }
-
-  @Test
-  void testCreateEncumbranceFromInactiveBudget() throws MalformedURLException {
-
-    verifyCollectionQuantity(TRANSACTION_ENDPOINT, TRANSACTION.getInitialQuantity(), TRANSACTION_TENANT_HEADER);
-
-    // prepare budget
-    Budget sampleBudget = (new JsonObject(getFile(BUDGET_SAMPLE))).mapTo(Budget.class);
-    sampleBudget.setBudgetStatus(Budget.BudgetStatus.INACTIVE);
-    putData(BUDGET.getEndpointWithId(), sampleBudget.getId(), JsonObject.mapFrom(sampleBudget).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-      .statusCode(204);
-
-    String orderId = UUID.randomUUID().toString();
-    createOrderSummary(orderId, 1);
-    JsonObject jsonTx = new JsonObject(getFile(ENCUMBRANCE_SAMPLE));
-    jsonTx.remove("id");
-
-    Transaction encumbrance = jsonTx.mapTo(Transaction.class);
-    encumbrance.setAmount((double) Integer.MAX_VALUE);
-    encumbrance.getEncumbrance().setSourcePurchaseOrderId(orderId);
-
-    // create Encumbrance
-    postData(TRANSACTION_ENDPOINT, JsonObject.mapFrom(encumbrance).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-      .statusCode(400)
-      .body(containsString(BUDGET_IS_INACTIVE));
-
-  }
-
-  @Test
-  void testCreateEncumbranceWithoutSummary() throws MalformedURLException {
-
-    String orderId = UUID.randomUUID().toString();
-
-    JsonObject jsonTx = new JsonObject(getFile(ENCUMBRANCE_SAMPLE));
-    jsonTx.remove("id");
-    Transaction encumbrance = jsonTx.mapTo(Transaction.class);
-
-    encumbrance.getEncumbrance().setSourcePurchaseOrderId(orderId);
-
-    String transactionSample = JsonObject.mapFrom(encumbrance).encodePrettily();
-
-    postData(TRANSACTION_ENDPOINT, transactionSample, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(400);
-
-  }
-
-  @Test
-  void testCreateEncumbranceWithoutBudget() throws MalformedURLException {
-
-    String orderId = UUID.randomUUID().toString();
-    createOrderSummary(orderId, 2);
-    JsonObject jsonTx = new JsonObject(getFile(ENCUMBRANCE_SAMPLE));
-    jsonTx.remove("id");
-
-    Transaction encumbrance = jsonTx.mapTo(Transaction.class);
-    encumbrance.setFiscalYearId(UUID.randomUUID().toString());
-
-    encumbrance.getEncumbrance().setSourcePurchaseOrderId(orderId);
-
-    String transactionSample = JsonObject.mapFrom(encumbrance).encodePrettily();
-
-    postData(TRANSACTION_ENDPOINT, transactionSample, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(400).body(containsString(BUDGET_NOT_FOUND_FOR_TRANSACTION));
-
-  }
-
-  protected void createOrderSummary(String orderId, int encumbranceNumber) throws MalformedURLException {
-    OrderTransactionSummary summary = new OrderTransactionSummary().withId(orderId).withNumTransactions(encumbranceNumber);
-    postData(ORDER_TRANSACTION_SUMMARIES_ENDPOINT, JsonObject.mapFrom(summary)
-      .encodePrettily(), TRANSACTION_TENANT_HEADER);
-  }
-
-  @Test
-  void testCreateEncumbranceWithMissedRequiredFields() throws MalformedURLException {
-
-    JsonObject jsonTx = new JsonObject(getFile(ENCUMBRANCE_SAMPLE));
-
-    Transaction encumbrance = jsonTx.mapTo(Transaction.class);
-
-    encumbrance.setEncumbrance(null);
-    encumbrance.setFromFundId(null);
-
-    String transactionSample = JsonObject.mapFrom(encumbrance).encodePrettily();
-
-    Errors errors = postData(TRANSACTION_ENDPOINT, transactionSample, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(422).extract().as(Errors.class);
-    assertThat(errors.getErrors(), hasSize(2));
-
-  }
-
-  @Test
-  void testCreateEncumbrancesDuplicateInTemporaryTable() throws MalformedURLException {
-
-    String orderId = UUID.randomUUID().toString();
-
-    createOrderSummary(orderId, 2);
-    JsonObject jsonTx = new JsonObject(getFile(ENCUMBRANCE_SAMPLE));
-    jsonTx.remove("id");
-    Transaction encumbrance = jsonTx.mapTo(Transaction.class);
-
-    encumbrance.getEncumbrance().setSourcePurchaseOrderId(orderId);
-
-    String transactionSample = JsonObject.mapFrom(encumbrance).encodePrettily();
-
-    String encumbranceId = postData(TRANSACTION_ENDPOINT, transactionSample, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201).extract().as(Transaction.class).getId();
-
-    // encumbrance do not appear in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), encumbranceId, TRANSACTION_TENANT_HEADER).then().statusCode(404);
-
-    // create encumbrance again
-    postData(TRANSACTION_ENDPOINT, transactionSample, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201);
-
-    // encumbrance do not appear in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), encumbranceId, TRANSACTION_TENANT_HEADER).then().statusCode(404);
-
-  }
-
-  @Test
-  void testCreateEncumbrancesDuplicateInTransactionTable() throws MalformedURLException {
-
-    String orderId = UUID.randomUUID().toString();
-    createOrderSummary(orderId,  2);
-
-    JsonObject jsonTx = new JsonObject(getFile(ENCUMBRANCE_SAMPLE));
-    jsonTx.remove("id");
-
-    Transaction encumbrance1 = jsonTx.mapTo(Transaction.class);
-    encumbrance1.getEncumbrance().setSourcePurchaseOrderId(orderId);
-
-    Transaction encumbrance2 = jsonTx.mapTo(Transaction.class);
-    encumbrance2.getEncumbrance().setSourcePurchaseOrderId(orderId);
-    encumbrance2.getEncumbrance().setSourcePoLineId(UUID.randomUUID().toString());
-
-    String transactionSample1 = JsonObject.mapFrom(encumbrance1).encodePrettily();
-
-    // create 1st Encumbrance, expected number is 2
-    String encumbrance1Id = postData(TRANSACTION_ENDPOINT, transactionSample1, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201)
-      .extract()
-      .as(Transaction.class).getId();
-
-    // encumbrance do not appear in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), encumbrance1Id, TRANSACTION_TENANT_HEADER).then().statusCode(404);
-
-
-    String transactionSample2 = JsonObject.mapFrom(encumbrance2).encodePrettily();
-
-    // create 2nd Encumbrance
-    String encumbrance2Id = postData(TRANSACTION_ENDPOINT, transactionSample2, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201)
-      .extract()
-      .as(Transaction.class).getId();
-
-    // 2 encumbrances appear in transaction table, temp transactions deleted
-    getDataById(TRANSACTION.getEndpointWithId(), encumbrance1Id, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-    getDataById(TRANSACTION.getEndpointWithId(), encumbrance2Id, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-
-    encumbrance1Id = postData(TRANSACTION_ENDPOINT, transactionSample1, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201)
-      .extract()
-      .as(Transaction.class).getId();
-
-    // encumbrance do not appear in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), encumbrance1Id, TRANSACTION_TENANT_HEADER).then().statusCode(404);
-
-    // create 2nd Encumbrance
-    postData(TRANSACTION_ENDPOINT, transactionSample2, TRANSACTION_TENANT_HEADER)
-      .then()
-      .statusCode(201);
-
-  }
-
-  @Test
-  void testUpdateEncumbranceAllOrNothing() throws MalformedURLException {
-
-    String orderId = UUID.randomUUID().toString();
-    createOrderSummary(orderId, 2);
-
-    JsonObject jsonTx = new JsonObject(getFile(ENCUMBRANCE_SAMPLE));
-    jsonTx.remove("id");
-    Transaction encumbrance1 = jsonTx.mapTo(Transaction.class);
-    encumbrance1.getEncumbrance().setSourcePurchaseOrderId(orderId);
-
-    Transaction encumbrance2 = jsonTx.mapTo(Transaction.class);
-    encumbrance2.getEncumbrance().setSourcePurchaseOrderId(orderId);
-    encumbrance2.getEncumbrance().setSourcePoLineId(UUID.randomUUID().toString());
-
-    String fY = encumbrance1.getFiscalYearId();
-    String fromFundId = encumbrance1.getFromFundId();
-
-    String transactionSample = JsonObject.mapFrom(encumbrance1).encodePrettily();
-    // prepare budget queries
-    String fromBudgetEndpointWithQueryParams = String.format(BUDGETS_QUERY, fY, fromFundId);
-
-    // create 1st Encumbrance, expected number is 2
-    String encumbrance1Id = postData(TRANSACTION_ENDPOINT, transactionSample, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201)
-      .extract()
-      .as(Transaction.class).getId();
-
-    // encumbrance do not appear in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), encumbrance1Id, TRANSACTION_TENANT_HEADER).then().statusCode(404);
-
-    transactionSample = JsonObject.mapFrom(encumbrance2).encodePrettily();
-
-    // create 2nd Encumbrance
-    String encumbrance2Id = postData(TRANSACTION_ENDPOINT, transactionSample, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201)
-      .extract()
-      .as(Transaction.class).getId();
-
-    // 2 encumbrances appear in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), encumbrance1Id, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-    getDataById(TRANSACTION.getEndpointWithId(), encumbrance2Id, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-
-    Budget fromBudgetBefore = getBudgetAndValidate(fromBudgetEndpointWithQueryParams);
-
-    double releasedAmount = encumbrance1.getAmount();
-    double amountAwaitingPaymentDif = 5.5;
-    encumbrance1.getEncumbrance().setStatus(Encumbrance.Status.RELEASED);
-    encumbrance2.setAmount(100d);
-    encumbrance2.getEncumbrance().setStatus(Encumbrance.Status.UNRELEASED);
-    encumbrance2.getEncumbrance().setAmountAwaitingPayment(sumValues(encumbrance2.getEncumbrance().getAmountAwaitingPayment(), 5.5));
-
-    // First encumbrance update, save to temp table, changes won't get to transaction table
-    putData(TRANSACTION.getEndpointWithId(), encumbrance1Id, JsonObject.mapFrom(encumbrance1).encodePrettily(), TRANSACTION_TENANT_HEADER).then().statusCode(204);
-    Transaction transaction1FromStorage = getDataById(TRANSACTION.getEndpointWithId(), encumbrance1Id, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-    assertEquals(Encumbrance.Status.UNRELEASED, transaction1FromStorage.getEncumbrance().getStatus());
-    assertEquals(transaction1FromStorage.getAmount(), releasedAmount);
-
-    // Second encumbrance update, changes for two encumbrances will get to transaction table
-    putData(TRANSACTION.getEndpointWithId(), encumbrance2Id, JsonObject.mapFrom(encumbrance2).encodePrettily(), TRANSACTION_TENANT_HEADER).then().statusCode(204);
-    transaction1FromStorage = getDataById(TRANSACTION.getEndpointWithId(), encumbrance1Id, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-    Transaction transaction2FromStorage = getDataById(TRANSACTION.getEndpointWithId(), encumbrance2Id, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-
-    assertEquals(Encumbrance.Status.RELEASED, transaction1FromStorage.getEncumbrance().getStatus());
-    assertEquals(0d, transaction1FromStorage.getAmount());
-    assertEquals(transaction2FromStorage.getEncumbrance().getAmountAwaitingPayment(), encumbrance2.getEncumbrance().getAmountAwaitingPayment());
-    double expectedAmount = subtractValues(encumbrance2.getEncumbrance().getInitialAmountEncumbered(), encumbrance2.getEncumbrance().getAmountAwaitingPayment());
-    expectedAmount = subtractValues(expectedAmount, encumbrance2.getEncumbrance().getAmountExpended());
-    assertEquals(expectedAmount, transaction2FromStorage.getAmount());
-
-    Budget fromBudgetAfterUpdate = getBudgetAndValidate(fromBudgetEndpointWithQueryParams);
-
-    double expectedBudgetsEncumbered = subtractValues(fromBudgetBefore.getEncumbered(), releasedAmount);
-    expectedBudgetsEncumbered = subtractValues(expectedBudgetsEncumbered, amountAwaitingPaymentDif);
-    double expectedBudgetsAvailable = sumValues(fromBudgetBefore.getAvailable(), releasedAmount);
-    double expectedBudgetsUnavailable = subtractValues(fromBudgetBefore.getUnavailable(), releasedAmount);
-    expectedBudgetsUnavailable = expectedBudgetsUnavailable < 0 ? 0 : expectedBudgetsUnavailable;
-    double expectedAwaitingPayment = sumValues(fromBudgetBefore.getAwaitingPayment(), amountAwaitingPaymentDif);
-
-    assertEquals(expectedBudgetsEncumbered, fromBudgetAfterUpdate.getEncumbered());
-    assertEquals(expectedBudgetsAvailable , fromBudgetAfterUpdate.getAvailable());
-    assertEquals(expectedBudgetsUnavailable, fromBudgetAfterUpdate.getUnavailable());
-    assertEquals(expectedAwaitingPayment, fromBudgetAfterUpdate.getAwaitingPayment());
-
-  }
-
-  @Test
-  void testUpdateAlreadyReleasedEncumbranceBudgetNotUpdated() throws MalformedURLException {
-
-    String orderId = UUID.randomUUID().toString();
-      createOrderSummary(orderId, 1);
-
-    JsonObject jsonTx = new JsonObject(getFile(ENCUMBRANCE_SAMPLE));
-    jsonTx.remove("id");
-    Transaction encumbrance = jsonTx.mapTo(Transaction.class);
-    encumbrance.getEncumbrance().setSourcePurchaseOrderId(orderId);
-    encumbrance.getEncumbrance().setStatus(Encumbrance.Status.RELEASED);
-    encumbrance.setAmount(0d);
-    encumbrance.getEncumbrance().setAmountAwaitingPayment(10d);
-
-    String fY = encumbrance.getFiscalYearId();
-    String fromFundId = encumbrance.getFromFundId();
-
-    String transactionSample = JsonObject.mapFrom(encumbrance).encodePrettily();
-    // prepare budget queries
-    String fromBudgetEndpointWithQueryParams = String.format(BUDGETS_QUERY, fY, fromFundId);
-
-    // create 1st Encumbrance, expected number is 2
-    String encumbrance1Id = postData(TRANSACTION_ENDPOINT, transactionSample, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201)
-      .extract()
-      .as(Transaction.class).getId();
-
-    // encumbrance appearS in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), encumbrance1Id, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-
-    Budget fromBudgetBefore = getBudgetAndValidate(fromBudgetEndpointWithQueryParams);
-
-    encumbrance.getEncumbrance().setAmountAwaitingPayment(5d);
-    encumbrance.setAmount(2d);
-
-    putData(TRANSACTION.getEndpointWithId(), encumbrance1Id, JsonObject.mapFrom(encumbrance).encodePrettily(), TRANSACTION_TENANT_HEADER).then().statusCode(204);
-    Transaction transaction1FromStorage = getDataById(TRANSACTION.getEndpointWithId(), encumbrance1Id, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-    assertEquals(5d, transaction1FromStorage.getEncumbrance().getAmountAwaitingPayment());
-    assertEquals(2d, transaction1FromStorage.getAmount());
-    Budget fromBudgetAfterUpdate = getBudgetAndValidate(fromBudgetEndpointWithQueryParams);
-
-    assertEquals(fromBudgetBefore.getEncumbered(), fromBudgetAfterUpdate.getEncumbered());
-    assertEquals(fromBudgetBefore.getAvailable() , fromBudgetAfterUpdate.getAvailable());
-    assertEquals(fromBudgetBefore.getUnavailable(), fromBudgetAfterUpdate.getUnavailable());
-    assertEquals(fromBudgetBefore.getAwaitingPayment(), fromBudgetAfterUpdate.getAwaitingPayment());
-
-  }
-
-
-  @Test
-  void testUpdateEncumbranceNotFound() throws MalformedURLException {
-
-    String orderId = UUID.randomUUID().toString();
-    createOrderSummary(orderId, 2);
-
-    JsonObject jsonTx = new JsonObject(getFile(ENCUMBRANCE_SAMPLE));
-    jsonTx.remove("id");
-    Transaction encumbrance = jsonTx.mapTo(Transaction.class);
-    encumbrance.getEncumbrance().setSourcePurchaseOrderId(orderId);
-
-    // Try to update non-existent transaction
-    putData(TRANSACTION.getEndpointWithId(), UUID.randomUUID().toString(), JsonObject.mapFrom(encumbrance).encodePrettily(), TRANSACTION_TENANT_HEADER).then().statusCode(404);
-
-  }
-
-  @Test
-  void tesPostEncumbranceUnavailableMustNotIncludeOverEncumberedAmounts() throws MalformedURLException {
-    FiscalYear fiscalYear = new JsonObject(getFile(FISCAL_YEAR.getPathToSampleFile())).mapTo(FiscalYear.class).withId(null)
-      .withCode("TEST2020");
-    String fiscalYearId =  postData(FISCAL_YEAR.getEndpoint(), JsonObject.mapFrom(fiscalYear).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201).extract().as(FiscalYear.class).getId();
-
-    Ledger ledger = new JsonObject(getFile(LEDGER.getPathToSampleFile())).mapTo(Ledger.class)
-      .withId(null)
-      .withFiscalYearOneId(fiscalYearId)
-      .withCode("test")
-      .withName("test")
-      .withRestrictEncumbrance(true);
-    String ledgerId =  postData(LEDGER.getEndpoint(), JsonObject.mapFrom(ledger).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201).extract().as(Ledger.class).getId();
-
-    Fund fund = new JsonObject(getFile(FUND.getPathToSampleFile())).mapTo(Fund.class).withId(null).withLedgerId(ledgerId).withCode("test");
-    String fundId =  postData(FUND.getEndpoint(), JsonObject.mapFrom(fund).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201).extract().as(Fund.class).getId();
-
-    final double allocated = 100d;
-    final double available = 70d;
-    final double unavailable = 30d;
-    final double overEncumbrance = 150d;
-    Budget budget = new JsonObject(getFile(BUDGET.getPathToSampleFile())).mapTo(Budget.class)
-      .withId(null)
-      .withName("test")
-      .withFundId(fundId)
-      .withFiscalYearId(fiscalYearId)
-      .withBudgetStatus(Budget.BudgetStatus.ACTIVE)
-      .withAllocated(allocated)
-      .withAvailable(available)
-      .withUnavailable(unavailable)
-      .withAllowableExpenditure(overEncumbrance)
-      .withOverEncumbrance(0d);
-
-    String ledgerFYEndpointWithQueryParams = String.format(LEDGER_FYS_ENDPOINT, ledgerId, fiscalYearId);
-
-    LedgerFY ledgerFYBefore = getLedgerFYAndValidate(ledgerFYEndpointWithQueryParams);
-
-    String budgetId =  postData(BUDGET.getEndpoint(), JsonObject.mapFrom(budget).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201).extract().as(Budget.class).getId();
-
-    String orderId = UUID.randomUUID().toString();
-    createOrderSummary(orderId, 1);
-
-    JsonObject jsonTx = new JsonObject(getFile(ENCUMBRANCE_SAMPLE));
-    jsonTx.remove("id");
-    Transaction encumbrance = jsonTx.mapTo(Transaction.class);
-
-    encumbrance.getEncumbrance().setSourcePurchaseOrderId(orderId);
-    encumbrance.setFromFundId(fundId);
-    encumbrance.setFiscalYearId(fiscalYearId);
-    final double transactionAmount = 100d;
-    encumbrance.setAmount(transactionAmount);
-
-    String transactionSample = JsonObject.mapFrom(encumbrance).encodePrettily();
-
-    String encumbranceId = postData(TRANSACTION_ENDPOINT, transactionSample, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201).extract().as(Transaction.class).getId();
-
-    // encumbrance do not appear in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), encumbranceId, TRANSACTION_TENANT_HEADER).then().statusCode(200);
-
-    Budget budgetAfter = getDataById(BUDGET.getEndpointWithId(), budgetId, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(200).extract().as(Budget.class);
-
-    Double expectedOverEncumbrance = subtractValues(transactionAmount, available);
-    assertEquals(allocated, budgetAfter.getUnavailable());
-    assertEquals(expectedOverEncumbrance, budgetAfter.getOverEncumbrance());
-
-    LedgerFY ledgerFYAfter = getLedgerFYAndValidate(ledgerFYEndpointWithQueryParams);
-    assertEquals(available, subtractValues(ledgerFYAfter.getUnavailable(), ledgerFYBefore.getUnavailable()));
-  }
-
   private Ledger getLedgerAndValidate(String endpoint) throws MalformedURLException {
     return getData(endpoint, TRANSACTION_TENANT_HEADER).then()
       .statusCode(200)
@@ -885,14 +352,6 @@ class TransactionTest extends TestBase {
       .body(BUDGETS, hasSize(1))
       .extract()
       .as(BudgetCollection.class).getBudgets().get(0);
-  }
-
-  protected double subtractValues(double d1, double d2) {
-    return BigDecimal.valueOf(d1).subtract(BigDecimal.valueOf(d2)).doubleValue();
-  }
-
-  private double sumValues(double d1, double d2) {
-    return BigDecimal.valueOf(d1).add(BigDecimal.valueOf(d2)).doubleValue();
   }
 
 }
